switch telegram_chat_id and alert_id column data types from in to bigint

+++ src/schema.py
@@ -12,8 +12,8 @@ Base = declarative_base()
 
 class Alert(Base):
     __tablename__ = "active_alerts"
-    alert_id = Column(Integer, primary_key=True)
-    telegram_chat_id = Column(Integer, nullable=False)
+    alert_id = Column(BigInteger, primary_key=True)
+    telegram_chat_id = Column(BigInteger, nullable=False)
     gas_threshold_gwei = Column(Integer, nullable=False)
     cooldown_seconds = Column(Integer, nullable=False)
     cooldown_expired_timestamp = Column(BigInteger, nullable=False)
